requirements.txt
python-telegram-bot==12.0.0b1
gspread
oauth2client